technical_indicator/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
pydantic==2.8.2
pydantic-settings==2.4.0
numpy==2.1.1
pandas==2.2.2
pyyaml==6.0.2
# Si tu utilises TA-Lib natif, il faut libstdc++/compil — sinon garde tes modules maison
# ta==0.11.0   # (facultatif si tu t’appuies sur tes propres fichiers d’indicateurs)
